update advanced reboot test for kvm

+++ ansible/roles/test/files/ptftests/advanced-reboot.py
@@ -126,6 +126,7 @@ class ReloadTest(BaseTest):
         self.sad_handle = None
         self.process_id = str(os.getpid())
         self.test_params = testutils.test_params_get()
+        self.check_param('kvm_test', False, required = False)
         self.check_param('verbose', False, required=False)
         self.check_param('dut_username', '', required=True)
         self.check_param('dut_password', '', required=True)
@@ -155,7 +156,6 @@ class ReloadTest(BaseTest):
         self.check_param('vnet', False, required = False)
         self.check_param('vnet_pkts', None, required = False)
         self.check_param('target_version', '', required = False)
-        self.check_param('kvm_test', True, required = False)
         if not self.test_params['preboot_oper'] or self.test_params['preboot_oper'] == 'None':
             self.test_params['preboot_oper'] = None
         if not self.test_params['inboot_oper'] or self.test_params['inboot_oper'] == 'None':
@@ -493,12 +493,12 @@ class ReloadTest(BaseTest):
         self.default_ip_range = self.test_params['default_ip_range']
 
         self.limit = datetime.timedelta(seconds=self.test_params['reboot_limit_in_seconds'])
+        self.kvm_test = self.test_params['kvm_test']
         self.reboot_type = self.test_params['reboot_type']
         if self.reboot_type not in ['fast-reboot', 'warm-reboot', 'warm-reboot -f']:
             raise ValueError('Not supported reboot_type %s' % self.reboot_type)
         self.dut_mac = self.test_params['dut_mac']
 
-        self.kvm_test = self.test_params['kvm_test']
         if self.kvm_test:
             self.log("This test is for KVM platform")
 
@@ -757,8 +757,6 @@ class ReloadTest(BaseTest):
         self.reboot_start = None
         self.no_routing_start = None
         self.no_routing_stop = None
-        self.no_control_start = None
-        self.no_control_stop = None
         self.no_cp_replies = None
         self.upper_replies = []
         self.routing_always = False
@@ -795,8 +793,9 @@ class ReloadTest(BaseTest):
             self.do_inboot_oper()
         self.reboot_start = datetime.datetime.now()
         self.log("Dut reboots: reboot start %s" % str(self.reboot_start))
-        self.no_control_start = self.cpu_state.get_state_time('down')
-        self.log("control plane down starts %s" % str(self.no_control_start))
+        if self.kvm_test:
+            self.no_control_start = self.cpu_state.get_state_time('down')
+            self.log("control plane down starts %s" % str(self.no_control_start))
 
     def handle_fast_reboot_health_check(self):
         if not self.kvm_test: # The data plane is expected to go down with control plane in KVM tests
@@ -851,11 +850,6 @@ class ReloadTest(BaseTest):
 
         self.save_sniffed_packets()
 
-        self.log("Wait until control plane up")
-        async_cpu_up = self.pool.apply_async(self.wait_until_cpu_port_up)
-        self.no_control_stop = self.cpu_state.get_state_time('up')
-        self.log("Control plane down stops %s" % str(self.no_control_stop))
-
         examine_start = datetime.datetime.now()
         self.log("Packet flow examine started %s after the reboot" % str(examine_start - self.reboot_start))
         self.examine_flow()
@@ -870,29 +864,27 @@ class ReloadTest(BaseTest):
             self.no_routing_start = self.reboot_start
             self.no_routing_stop  = self.reboot_start
 
-    def handle_warm_reboot_health_check_kvm(self):
+    def handle_advanced_reboot_health_check_kvm(self):
         self.log("Wait until data plane stops")
         async_forward_stop = self.pool.apply_async(self.check_forwarding_stop)
 
         self.log("Wait until control plane up")
         async_cpu_up = self.pool.apply_async(self.wait_until_cpu_port_up)
-        self.no_control_stop = self.cpu_state.get_state_time('up')
-        self.log("Control plane down stops %s" % str(self.no_control_stop))
+
+        try:
+            self.no_routing_start, _ = async_forward_stop.get(timeout=self.task_timeout)
+            self.log("Data plane was stopped, Waiting until it's up. Stop time: %s" % str(self.no_routing_start))
+        except TimeoutError:
+            self.log("Data plane never stop")
 
         try:
             async_cpu_up.get(timeout=self.task_timeout)
+            self.no_control_stop = self.cpu_state.get_state_time('up')
+            self.log("Control plane down stops %s" % str(self.no_control_stop))
         except TimeoutError as e:
             self.log("DUT hasn't bootup in %d seconds" % self.task_timeout)
             self.fails['dut'].add("DUT hasn't booted up in %d seconds" % self.task_timeout)
             raise
-
-        try:
-            self.no_routing_start, self.upper_replies = async_forward_stop.get(timeout=self.task_timeout)
-            self.log("Data plane was stopped, Waiting until it's up. Stop time: %s" % str(self.no_routing_start))
-        except TimeoutError:
-            self.log("Data plane never stop")
-            self.routing_always = True
-            self.upper_replies = [self.nr_vl_pkts]
 
         if self.no_routing_start is not None:
             self.no_routing_stop, _ = self.timeout(self.check_forwarding_resume,
@@ -923,16 +915,11 @@ class ReloadTest(BaseTest):
         self.log("Data plane works again. Start time: %s" % str(self.no_routing_stop))
         self.log("")
 
-        if self.reboot_type == 'fast-reboot':
-            self.no_cp_replies = self.extract_no_cpu_replies(self.upper_replies)
-
         if self.no_routing_stop - self.no_routing_start > self.limit:
             self.fails['dut'].add("Longest downtime period must be less then %s seconds. It was %s" \
                     % (self.test_params['reboot_limit_in_seconds'], str(self.no_routing_stop - self.no_routing_start)))
         if self.no_routing_stop - self.reboot_start > datetime.timedelta(seconds=self.test_params['graceful_limit']):
             self.fails['dut'].add("%s cycle must be less than graceful limit %s seconds" % (self.reboot_type, self.test_params['graceful_limit']))
-        if self.reboot_type == 'fast-reboot' and self.no_cp_replies < 0.95 * self.nr_vl_pkts:
-            self.fails['dut'].add("Dataplane didn't route to all servers, when control-plane was down: %d vs %d" % (self.no_cp_replies, self.nr_vl_pkts))
 
         if 'warm-reboot' in self.reboot_type:
             if self.total_disrupt_time > self.limit.total_seconds():
@@ -950,6 +937,11 @@ class ReloadTest(BaseTest):
             else:
                 # verify there are no interface flaps after warm boot
                 self.neigh_lag_status_check()
+
+        if self.reboot_type == 'fast-reboot' and not self.kvm_test:
+            self.no_cp_replies = self.extract_no_cpu_replies(self.upper_replies)
+            if self.no_cp_replies < 0.95 * self.nr_vl_pkts:
+                self.fails['dut'].add("Dataplane didn't route to all servers, when control-plane was down: %d vs %d" % (self.no_cp_replies, self.nr_vl_pkts))
 
     def handle_post_reboot_test_reports(self):
         # Stop watching DUT
@@ -992,7 +984,8 @@ class ReloadTest(BaseTest):
 
         if self.no_routing_stop:
             self.log("Longest downtime period was %s" % str(self.no_routing_stop - self.no_routing_start))
-            self.log("Control plane downtime period was %s" % str(self.no_control_stop - self.no_control_start))
+            if self.kvm_test:
+                self.log("Control plane downtime period was %s" % str(self.no_control_stop - self.no_control_start))
             reboot_time = "0:00:00" if self.routing_always else str(self.no_routing_stop - self.reboot_start)
             self.log("Reboot time was %s" % reboot_time)
             self.log("Expected downtime is less then %s" % self.limit)
@@ -1042,10 +1035,14 @@ class ReloadTest(BaseTest):
             thr.start()
 
             self.wait_until_reboot()
-            if self.reboot_type == 'fast-reboot':
-                self.handle_fast_reboot_health_check()
-            if 'warm-reboot' in self.reboot_type:
-                self.handle_warm_reboot_health_check()
+            if self.kvm_test:
+                self.handle_advanced_reboot_health_check_kvm()
+            else:
+                if self.reboot_type == 'fast-reboot':
+                    self.handle_fast_reboot_health_check()
+                if 'warm-reboot' in self.reboot_type:
+                    self.handle_warm_reboot_health_check()
+
             self.handle_post_reboot_health_check()
 
             # Check sonic version after reboot
